+++ Dockerfile
@@ -1,0 +1,16 @@
+FROM node:lts-alpine
+
+# make the 'app' folder the current working directory
+WORKDIR /app
+
+# copy both 'package.json' and 'package-lock.json' (if available)
+COPY package*.json ./
+
+# install project dependencies
+RUN npm install
+
+# copy project files and folders to the current working directory (i.e. 'app' folder)
+COPY . .
+
+# No ports are required to be exposed for a discord bot. Just run the bot
+CMD ["node", "src/index.js"]
