--- commands/say.js
@@ -1,4 +1,0 @@
-exports.run = (client, message, args)=>{
-    var response = args.join();
-    message.channel.send(response);
-}
