+++ docker-compose.yml
@@ -1,0 +1,9 @@
+  
+version: '3'
+
+services:
+  discord:
+    build: .
+    volumes:
+      - .:/usr/src/app
+      - /usr/src/app/node_modules
