--- example_tokens.md
@@ -1,5 +1,0 @@
-#This is an example of the format for the 'tokens.json' file.  
-
-{
-    "token":"YOUR_UNIQUE_BOT_ID"
-}
