+++ src/commands/ping.js
@@ -1,3 +1,3 @@
 exports.run = (client, message, args) => { //Override the default args with the client arg from earlier.
-    message.channel.send("pong!").catch(console.error);
-}
+	message.channel.send("pong!").catch(console.error);
+};
