+++ src/commands/say.js
@@ -1,0 +1,4 @@
+exports.run = (client, message, args)=>{
+	var response = args.join();
+	message.channel.send(response);
+};
