+++ src/tokens.json.example
@@ -1,0 +1,3 @@
+{
+    "token":"YOUR_UNIQUE_BOT_ID"
+}
